refactor: split installed dashboard smoke helper
